refactor: move personal data from the form into a separate feature
