+++ src/features/personalForm/index.tsx
@@ -1,0 +1,1 @@
+export { PersonalForm } from './ui/PersonalForm';
