+++ src/features/personalForm/model/index.ts
@@ -1,0 +1,3 @@
+export * from './constants';
+export * from './useValidators';
+export * from './types';
